split parse-capi up
