--- packages/frontend/lib/parse-capi/big-bullets.ts
@@ -1,2 +1,0 @@
-export default (s: string) =>
-    s.replace(/•/g, '<span class="bullet">&bull;</span>');
