+++ packages/frontend/lib/parse-capi/extract-capi.ts
@@ -1,8 +1,4 @@
 import { string as curly } from 'curlyquotes';
-import get from 'lodash.get';
-
-import clean from './clean';
-import bigBullets from './big-bullets';
 import {
     getString,
     getNumber,
@@ -10,22 +6,16 @@ import {
     getBoolean,
     getArray,
 } from './validators';
+import clean from './clean';
 import { getSharingUrls } from './sharing-urls';
-import { pillarNames } from '@frontend/lib/pillars';
+import { findPillar } from './find-pillar';
+
+const bigBullets = (s: string) =>
+    s.replace(/•/g, '<span class="bullet">&bull;</span>');
 
 // tslint:disable:prefer-array-literal
 const apply = (input: string, ...fns: Array<(_: string) => string>): string => {
     return fns.reduce((acc, fn) => fn(acc), input);
-};
-
-const findPillar: (name: string) => Pillar | undefined = name => {
-    const pillar: string = name.toLowerCase();
-    // The pillar name is "arts" in CAPI, but "culture" everywhere else,
-    // therefore we perform this substitution here.
-    if (pillar === 'arts') {
-        return 'culture';
-    }
-    return pillarNames.find(_ => _ === pillar);
 };
 
 const getEditionValue: (name: string) => Edition = name => {
@@ -34,18 +24,16 @@ const getEditionValue: (name: string) => Edition = name => {
     return edition === undefined ? 'UK' : edition;
 };
 
-const getLink = (data: {}, { isPillar }: { isPillar: boolean }): LinkType => {
-    const title = getString(data, 'title');
-    return {
-        title,
-        longTitle: getString(data, 'longTitle') || title,
-        url: getString(data, 'url'),
-        pillar: isPillar ? findPillar(getString(data, 'title')) : undefined,
-        children: getArray<object>(data, 'children', []).map(
-            l => getLink(l, { isPillar: false }), // children are never pillars
-        ),
-        mobileOnly: false,
-    };
+const getTags: (data: any) => TagType[] = data => {
+    const tags = getArray<any>(data, 'tags.tags', []);
+    return tags.map(tag => {
+        return {
+            id: getNonEmptyString(tag, 'properties.id'),
+            type: getNonEmptyString(tag, 'properties.tagType'),
+            title: getString(tag, 'properties.webTitle', ''),
+            twitterHandle: getString(tag, 'properties.twitterHandle', ''),
+        };
+    });
 };
 
 const getAgeWarning = (
@@ -155,18 +143,6 @@ const getSubMetaSectionLinks: (
     return links;
 };
 
-const getTags: (data: any) => TagType[] = data => {
-    const tags = getArray<any>(data, 'tags.tags', []);
-    return tags.map(tag => {
-        return {
-            id: getNonEmptyString(tag, 'properties.id'),
-            type: getNonEmptyString(tag, 'properties.tagType'),
-            title: getString(tag, 'properties.webTitle', ''),
-            twitterHandle: getString(tag, 'properties.twitterHandle', ''),
-        };
-    });
-};
-
 const getSubMetaKeywordLinks: (
     data: {
         tags: TagType[];
@@ -196,7 +172,7 @@ const getSubMetaKeywordLinks: (
 // TODO really it would be nice if we passed just the data we needed and
 // didn't have to do the transforms/lookups below. (While preserving the
 // validation on types.)
-export const extractArticleMeta = (data: {}): CAPIType => {
+export const extract = (data: {}): CAPIType => {
     const webPublicationDate = new Date(
         getNumber(data, 'config.page.webPublicationDate'),
     );
@@ -277,114 +253,5 @@ export const extractArticleMeta = (data: {}): CAPIType => {
             ...sectionData,
         }),
         ...sectionData,
-    };
-};
-
-export const extractNavMeta = (data: {}): NavType => {
-    let pillars = getArray<any>(data, 'config.nav.pillars');
-
-    pillars = pillars.map(link => getLink(link, { isPillar: true }));
-
-    const subnav = get(data, 'config.nav.subNavSections');
-
-    return {
-        pillars,
-        otherLinks: {
-            url: '', // unused
-            title: 'More',
-            longTitle: 'More',
-            more: true,
-            children: getArray<object>(data, 'config.nav.otherLinks', []).map(
-                l => getLink(l, { isPillar: false }),
-            ),
-        },
-        brandExtensions: getArray<object>(
-            data,
-            'config.nav.brandExtensions',
-            [],
-        ).map(l => getLink(l, { isPillar: false })),
-        currentNavLink: getString(data, 'config.nav.currentNavLink.title', ''),
-        subNavSections: subnav
-            ? {
-                  parent: subnav.parent
-                      ? getLink(subnav.parent, { isPillar: false })
-                      : undefined,
-                  links: getArray<object>(subnav, 'links').map(l =>
-                      getLink(l, { isPillar: false }),
-                  ),
-              }
-            : undefined,
-        // TODO: replace with data from Frontend
-        readerRevenueLinks: {
-            header: {
-                subscribe:
-                    'https://support.theguardian.com/subscribe?INTCMP=header_support_subscribe&amp;acquisitionData=%7B%22source%22%3A%22GUARDIAN_WEB%22%2C%22componentType%22%3A%22ACQUISITIONS_HEADER%22%2C%22componentId%22%3A%22header_support_subscribe%22%2C%22referrerPageviewId%22%3A%22jnmzvcxd5o7u2p94r35e%22%2C%22referrerUrl%22%3A%22https%3A%2F%2Fwww.theguardian.com%2Fuk%22%7D',
-                support:
-                    'https://support.theguardian.com/?INTCMP=header_support&amp;acquisitionData=%7B%22source%22%3A%22GUARDIAN_WEB%22%2C%22componentType%22%3A%22ACQUISITIONS_HEADER%22%2C%22componentId%22%3A%22header_support%22%2C%22referrerPageviewId%22%3A%22jnmzvcxd5o7u2p94r35e%22%2C%22referrerUrl%22%3A%22https%3A%2F%2Fwww.theguardian.com%2Fuk%22%7D',
-            },
-            footer: {
-                subscribe:
-                    'https://support.theguardian.com/subscribe?INTCMP=footer_support_subscribe&amp;acquisitionData=%7B%22source%22%3A%22GUARDIAN_WEB%22%2C%22componentType%22%3A%22ACQUISITIONS_FOOTER%22%2C%22componentId%22%3A%22footer_support_subscribe%22%2C%22referrerPageviewId%22%3A%22jnmzvcxd5o7u2p94r35e%22%2C%22referrerUrl%22%3A%22https%3A%2F%2Fwww.theguardian.com%2Fuk%22%7D',
-                contribute:
-                    'https://support.theguardian.com/contribute?INTCMP=footer_support_contribute&amp;acquisitionData=%7B%22source%22%3A%22GUARDIAN_WEB%22%2C%22componentType%22%3A%22ACQUISITIONS_FOOTER%22%2C%22componentId%22%3A%22footer_support_contribute%22%2C%22referrerPageviewId%22%3A%22jnmzvcxd5o7u2p94r35e%22%2C%22referrerUrl%22%3A%22https%3A%2F%2Fwww.theguardian.com%2Fuk%22%7D',
-            },
-            sideMenu: {
-                subscribe:
-                    'https://support.theguardian.com/subscribe?INTCMP=side_menu_support_subscribe&amp;acquisitionData=%7B&quot;source&quot;:&quot;GUARDIAN_WEB&quot;,&quot;componentType&quot;:&quot;ACQUISITIONS_HEADER&quot;,&quot;componentId&quot;:&quot;side_menu_support_subscribe&quot;%7D',
-                contribute:
-                    'https://support.theguardian.com/contribute?INTCMP=side_menu_support_contribute&amp;acquisitionData=%7B&quot;source&quot;:&quot;GUARDIAN_WEB&quot;,&quot;componentType&quot;:&quot;ACQUISITIONS_HEADER&quot;,&quot;componentId&quot;:&quot;side_menu_support_contribute&quot;%7D',
-            },
-        },
-    };
-};
-
-export const extractConfigMeta = (data: {}): ConfigType => {
-    return {
-        ajaxUrl: getNonEmptyString(data, 'config.page.ajaxUrl'),
-    };
-};
-
-export interface GADataType {
-    pillar: string;
-    webTitle: string;
-    section: string;
-    contentType: string;
-    commissioningDesks: string;
-    contentId: string;
-    authorIds: string;
-    keywordIds: string;
-    toneIds: string;
-    seriesId: string;
-    isHosted: string;
-    edition: string;
-    beaconUrl: string;
-}
-
-// All GA fields should  fall back to default values -
-// we should not bring down the website if a trackable field is missing!
-export const extractGAMeta = (data: {}): GADataType => {
-    const edition = getString(data, 'config.page.edition', '').toLowerCase();
-
-    return {
-        webTitle: getString(data, 'config.page.webTitle', ''),
-        pillar: findPillar(getString(data, 'config.page.pillar', '')) || 'news',
-        section: getString(data, 'config.page.section', ''),
-        contentType: getString(data, 'config.page.contentType', '')
-            .toLowerCase()
-            .split(' ')
-            .join(''),
-        commissioningDesks: getString(
-            data,
-            'config.page.commissioningDesks',
-            '',
-        ),
-        contentId: getString(data, 'config.page.contentId', ''),
-        authorIds: getString(data, 'config.page.authorIds', ''),
-        keywordIds: getString(data, 'config.page.keywordIds', ''),
-        toneIds: getString(data, 'config.page.toneIds', ''),
-        seriesId: getString(data, 'config.page.seriesId', ''),
-        isHosted: getBoolean(data, 'config.page.isHosted', false).toString(),
-        edition: edition === 'int' ? 'international' : edition,
-        beaconUrl: getString(data, 'config.page.beaconUrl', ''),
     };
 };
