+++ packages/frontend/lib/parse-capi/extract-config.ts
@@ -1,0 +1,5 @@
+import { getNonEmptyString } from './validators';
+
+export const extract = (data: {}): ConfigType => ({
+    ajaxUrl: getNonEmptyString(data, 'config.page.ajaxUrl'),
+});
