+++ packages/frontend/web/document.tsx
@@ -5,7 +5,7 @@ import { renderToString } from 'react-dom/server';
 import htmlTemplate from './htmlTemplate';
 import Article from './pages/Article';
 import assets from '@frontend/lib/assets';
-import { GADataType } from '@frontend/lib/parse-capi';
+import { GADataType } from '@frontend/lib/parse-capi/extract-ga';
 
 interface Props {
     data: {
